Consolidate mlkem-nistp hybrids into one macro
